Commands handled
Update 2: all the commands and inputs are handled. the length of
commands are checked,as well as the IndexOutOfBoundExceptions.  When a
vertical or horizontal segment command is called, if the starting row
(or column) is greater than the ending row (or column) the inputs are
inverted so that the segment is still drawn.
The output is printed on the console, not on another txt file.